requirements.txt
pytesseract
tesseract
pillow
opencv-python
pyautogui